refactor: consolidate options tools onto options-analysis-server
Closes #159.

stock-price had accreted the entire options surface alongside the technicals
it owns, and two tools — get_option_contracts and price_vertical_spread —
existed on BOTH servers with identical bodies hitting identical REST routes.
An agent connected to both saw duplicate tool names with no principled way to
choose between them.

Six tools move to options_analysis; the two duplicates are deleted from
stock-price rather than moved, since the options-analysis copies already exist
and are behaviourally identical (same signature, same route, same forwarded
params — only docstring wording differed):

  moved     get_full_options_chain, get_unusual_calls, get_delta_adjusted_oi,
            get_gamma_wall_history, get_oi_change_analysis, get_gex_profile
  deleted   get_option_contracts, price_vertical_spread (dupes)

  stock-price       33 -> 25 tools
  options-analysis   5 -> 11 tools

Bodies are unchanged and still one rest_client call deep, so the REST tier and
its OpenAPI surface are untouched. No deploy work either: unlike the arbitrage
split, both services already exist in both projects.

The reference sweep the issue predicted turned up three tests in
test_support_tools_adapters.py calling the moved tools through
stock_price_server. Repointed to options_analysis; the existing rest_client
patch still covers them because both wrappers import the same module object.

Guards added so the duplication cannot return:
  * absence assertions that all eight are gone from stock_price_server
  * routing assertions that the six movers stay one rest_client call deep
  * a general check that no get_/price_/analyze_ tool name is served by two
    wrappers at once — the class of defect this issue was about
  * ci_wrapper_smoke floor for options-analysis raised 1 -> 11

readme's server table updated: stock-price no longer claims options chain and
spread pricing; options-analysis now describes the full surface.

1143 backend tests pass; wrapper smoke 6/6 (stock-price 25, options-analysis
11); guards pass; OpenAPI unchanged at 105 routes.
